Fix dimensions for images with a selector
Problem
Image capture when using a selector will sometimes render the selected
element with different dimensions than the view dimensions when they
should match. We currently set the viewport to the dimensions of the
selected element when passed a selector and then render it inside that
viewport which can skew the element if dependant on its surrounding for
sizing.

Solution
Add a new prop `viewportSelector` which allows the viewport to be set
by a different selector. Then use the `boundingBox` of the element to
set the clip area to be screenshot resulting in the element being
rendered in the context of the full html. This also changes height and
width props to not both be required and lets the undefined dimension
be calculated.